fix pytest fixture function in package  and settings field missing
